automated_quiz_generator/requirements.txt
fastapi
uvicorn
pypdf
redis
pinecone
langchain-nomic
langchain-community
langchain_groq
python-dotenv
rank_bm25